Refactor Database type mapping and provide Default implementation to SchemaModel
